minor adjustments, change baseline

+++ src/simsettings.h
@@ -27,7 +27,7 @@
                                 type      value
   general */\
   XX(  debug,                   bool,     true)     /* print debug info to terminal */\
-  XX(  mujoco_timestep,         float,    0.002)    /* sim timestep in seconds - default 0.002 */\
+  XX(  mujoco_timestep,         double,   0.002)    /* sim timestep in seconds - default 0.002 */\
   XX(  curve_validation,        bool,     false)    /* save finger curve data for testing */\
   XX(  finger_stiffness,        double,       5)    /* mujoco finger joint spring stiffness, units unknown */\
   XX(  random_seed,             uint,         0)    /* random seed */\
